Update Collection.php
Propose removing class property $mainTable, as it's not required for anything. The only thing that was required for di compilation to work properly, was to define the type for the constructor parameter $mainTable within the phpdoc ... I have already tested out di compilation with m2.2.1 without the class property $mainTable, and it works.